Anime Desktop Project
This is the initial project commit. The entire code present in this commit is the skeleton of the project. It has the main functions the project should have. However, there are still a lot of things to do, such as implement missing features. Add a GUI and refactor the code since most of the code logic is in the main.