Refactoring the API calls
Modified the API calls that were previously only in the main by creating different API caller classes for each website's API used for the application.
Created new packages to separate the models and the API callers.
Added a method in the main to take user input and make sure the input is valid.